+++ Source/RethinkDb.Driver/Ast/ReqlAst.cs
@@ -58,49 +58,102 @@ namespace RethinkDb.Driver.Ast
             return list;
         }
 
-
-        public static JObject buildOptarg(OptArgs opts)
+        protected internal static JObject buildOptarg(OptArgs opts)
         {
             var dict = opts.ToDictionary(kvp => kvp.Key, kvp => kvp.Value.Build());
             return JObject.FromObject(dict);
         }
 
-        public async virtual Task<dynamic> runAsync<T>(Connection conn, object globalOpts = null)
+
+        
+        #region DYNAMIC LANGUAGE RUNTIME (DLR) RUNNERS
+
+        /// <summary>
+        /// Runs the query on the connection. If you know the response type
+        /// of your query, T (SUCCESS_ATOM) or Cursor[T] (SUCCESS_SEQUENCE or SUCCESS_PARTIAL)
+        /// it's recommended to use `.runAtom` or `.runCursor` helpers
+        /// as they offer a slight edge in performance since both bypass the 
+        /// dynamic language runtime execution engine.
+        /// </summary>
+        /// <returns>Returns T or Cursor[T]</returns>
+        public virtual Task<dynamic> runAsync<T>(Connection conn, object globalOpts = null)
         {
-            return await conn.runAsync<T>(this, globalOpts);
+            return conn.RunAsync<T>(this, globalOpts);
         }
 
-        public async virtual Task<dynamic> runAsync(Connection conn, object globalOpts = null)
+        /// <summary>
+        /// Runs the query on the connection. If you know the response type
+        /// of your query, T (SUCCESS_ATOM) or Cursor[T] (SUCCESS_SEQUENCE or SUCCESS_PARTIAL)
+        /// it's recommended to use `.runAtom` or `.runCursor` helpers
+        /// as they offer a slight edge in performance since both bypass the 
+        /// dynamic language runtime execution engine.
+        /// </summary>
+        /// <returns>Returns T or Cursor[T]</returns>
+        public virtual Task<dynamic> runAsync(Connection conn, object globalOpts = null)
         {
-            return await runAsync<dynamic>(conn, globalOpts);
+            return runAsync<dynamic>(conn, globalOpts);
         }
 
+        /// <summary>
+        /// Runs the query on the connection. If you know the response type
+        /// of your query, T (SUCCESS_ATOM) or Cursor[T] (SUCCESS_SEQUENCE or SUCCESS_PARTIAL)
+        /// it's recommended to use `.runAtom` or `.runCursor` helpers
+        /// as they offer a slight edge in performance since both bypass the 
+        /// dynamic language runtime execution engine.
+        /// </summary>
+        /// <returns>Returns T or Cursor[T]</returns>
         public virtual dynamic run<T>(Connection conn, object globalOpts = null)
         {
-            return conn.runAsync<T>(this, globalOpts).RunSync();
+            return conn.RunAsync<T>(this, globalOpts).WaitSync();
         }
 
+        /// <summary>
+        /// Runs the query on the connection. If you know the response type
+        /// of your query, T (SUCCESS_ATOM) or Cursor[T] (SUCCESS_SEQUENCE or SUCCESS_PARTIAL)
+        /// it's recommended to use `.runAtom` or `.runCursor` helpers
+        /// as they offer a slight edge in performance since both bypass the 
+        /// dynamic language runtime execution engine.
+        /// </summary>
+        /// <returns>Returns T or Cursor[T]</returns>
         public virtual dynamic run(Connection conn, object globalOpts = null)
         {
             return run<dynamic>(conn, globalOpts);
         }
 
+        #endregion
+
+
+
+
+
+
+        #region RESPONSE TYPED RUNNERS
+
         /// <summary>
-        /// Use this method if you're expecting a cursor from your query. This
-        /// method offers a slight edge in performance without the need for the
+        /// Executes a query with no expected response. Useful for fire-and-forget queries like insert, update.
+        /// </summary>
+        public void runNoReply(Connection conn, object globalOpts = null)
+        {
+            conn.RunNoReply(this, globalOpts);
+        }
+
+        /// <summary>
+        /// Use this method if you're expecting a cursor (SUCCESS_SEQUENCE or SUCCESS_PARTIAL) response
+        /// from your query. This method offers a slight edge in performance without the need for the
         /// dynamic language runtime like the run() method uses.
         /// </summary>
         /// <typeparam name="T"></typeparam>
         /// <param name="conn">connection</param>
         /// <param name="globalOpts">global anonymous type optional arguments</param>
         /// <returns>A Cursor</returns>
-        public async virtual Task<Cursor<T>> runCursorAsync<T>(Connection conn, object globalOpts = null)
+        public virtual Task<Cursor<T>> runCursorAsync<T>(Connection conn, object globalOpts = null)
         {
-            return await conn.runCursorAsync<T>(this, globalOpts);
+            return conn.RunCursorAsync<T>(this, globalOpts);
         }
+
         /// <summary>
-        /// Use this method if you're expecting a cursor from your query. This
-        /// method offers a slight edge in performance without the need for the
+        /// Use this method if you're expecting a cursor (SUCCESS_SEQUENCE or SUCCESS_PARTIAL) response
+        /// from your query. This method offers a slight edge in performance without the need for the
         /// dynamic language runtime like the run() method uses.
         /// </summary>
         /// <typeparam name="T"></typeparam>
@@ -109,65 +162,95 @@ namespace RethinkDb.Driver.Ast
         /// <returns>A Cursor</returns>
         public virtual Cursor<T> runCursor<T>(Connection conn, object globalOpts = null)
         {
-            return conn.runCursorAsync<T>(this, globalOpts).RunSync();
+            return conn.RunCursorAsync<T>(this, globalOpts).WaitSync();
         }
 
         /// <summary>
-        /// Helper shortcut for DML type of queries that returns # of inserts, deletes, errors.
+        /// Use this method if you're expecting SUCCESS_ATOM response from your query. This
+        /// method offers a slight edge in performance without the need for the
+        /// dynamic language runtime like the run() method uses.
         /// </summary>
-        public async virtual Task<Result> runResultAsync(Connection conn, object globalOpts = null)
+        public virtual Task<T> runAtomAsync<T>(Connection conn, object globalOpts = null)
         {
-            return await conn.runAsync<Result>(this, globalOpts) as Result;
+            return conn.RunAtomAsync<T>(this, globalOpts);
+        }
+
+        /// <summary>
+        /// Use this method if you're expecting SUCCESS_ATOM response from your query. This
+        /// method offers a slight edge in performance without the need for the
+        /// dynamic language runtime like the run() method uses.
+        /// </summary>
+        public virtual T runAtom<T>(Connection conn, object globalOpts = null)
+        {
+            return runAtomAsync<T>(conn, globalOpts).WaitSync();
+        }
+
+        #endregion
+
+
+        #region EXTRA RUN HELPERS
+
+        /// <summary>
+        /// Helper shortcut for DML type of queries that returns # of inserts, deletes, errors.
+        /// This method bypasses the dynamic language runtime for extra performance.
+        /// </summary>
+        public virtual Task<Result> runResultAsync(Connection conn, object globalOpts = null)
+        {
+            return conn.RunAtomAsync<Result>(this, globalOpts);
         }
 
         /// <summary>
         /// Helper shortcut for DML type of queries that returns # of inserts, deletes, errors.
+        /// This method bypasses the dynamic language runtime for extra performance.
         /// </summary>
         public virtual Result runResult(Connection conn, object globalOpts = null)
         {
-            return conn.runAsync<Result>(this, globalOpts).RunSync() as Result;
+            return runResultAsync(conn, globalOpts).WaitSync();
         }
 
         /// <summary>
         /// Helper shortcut for change feeds, use if your query is expecting an infinite changes() stream.
+        /// This method bypasses the dynamic language runtime for extra performance.
         /// </summary>
         /// <typeparam name="T">The document type of new/old value</typeparam>
-        public async virtual Task<Cursor<Change<T>>> runChangesAsync<T>(Connection conn, object globalOpts = null)
+        public virtual Task<Cursor<Change<T>>> runChangesAsync<T>(Connection conn, object globalOpts = null)
         {
-            return await conn.runCursorAsync<Change<T>>(this, globalOpts);
+            return conn.RunCursorAsync<Change<T>>(this, globalOpts);
         }
 
         /// <summary>
         /// Helper shortcut for change feeds, use if your query is expecting an infinite changes() stream.
+        /// This method bypasses the dynamic language runtime for extra performance.
         /// </summary>
         /// <typeparam name="T">The document type of new/old value</typeparam>
         public virtual Cursor<Change<T>> runChanges<T>(Connection conn, object globalOpts = null)
         {
-            return conn.runCursorAsync<Change<T>>(this, globalOpts).RunSync();
+            return conn.RunCursorAsync<Change<T>>(this, globalOpts).WaitSync();
         }
 
-        public void runNoReply(Connection conn, object globalOpts = null)
-        {
-            conn.runNoReply(this, globalOpts);
-        }
+        #endregion
+
 
         /// <summary>
         /// Helper shortcut for grouping queries.
+        /// This method bypasses the dynamic language runtime for extra performance.
         /// </summary>
         /// <typeparam name="TKey">The key type of how items are grouped</typeparam>
         /// <typeparam name="TItem">The type of items</typeparam>
         public async virtual Task<IEnumerable<GroupedResult<TKey, TItem>>> runGroupingAsync<TKey, TItem>(Connection conn, object globalOpts = null)
         {
-            return await runAsync<GroupedResult<TKey, TItem>>(conn, globalOpts);
+            var tsk = await runAtomAsync<GroupedResultSet<TKey, TItem>>(conn, globalOpts).UseInternalAwait();
+            return tsk;
         }
         /// <summary>
         /// Helper shortcut for grouping queries.
+        /// This method bypasses the dynamic language runtime for extra performance.
         /// </summary>
         /// <typeparam name="TKey">The key type of how items are grouped</typeparam>
         /// <typeparam name="TItem">The type of items</typeparam>
         public virtual IEnumerable<GroupedResult<TKey, TItem>> runGrouping<TKey, TItem>(Connection conn, object globalOpts = null)
         {
-            return runAsync<GroupedResult<TKey, TItem>>(conn, globalOpts).RunSync();
+            return runAtomAsync<GroupedResultSet<TKey, TItem>>(conn, globalOpts).WaitSync();
         }
 
     }