+++ HISTORY.md
@@ -1,11 +1,13 @@
 ## v0.0.7-alpha7
-* Asynchronicity: async/await run() implementations.
+* async/await run() implementations.
 * Database connection thread-safety.
 * EnsureSuccess() renamed to AssertNoErrors()
 * Assert: Deleted(), Inserted(), skipped(), replaced, etc.. helpers.
-* Better Reactive Extension (Rx) compatibility.
+* Better Reactive Extension (Rx) semantic compatibility.
 * Added Cursor.MoveNext(Timeout) for manual cursor movement.
 * Added new helper: runGrouping<TKey,TItem>()
+* Added new helper: runAtom<T>()
+* See project page documentation
 
 ## v0.0.7-alpha6
 * Added run helpers: runResult(), runChanges<T>()